pyproject.toml
[tool.poetry]
name = "secured"
version = "0.1.3"
description = ""
authors = ["Joao Paulo Euko"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
python-dotenv = "^1.0.1"
tomlkit = "^0.12.4"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_errors="True"
show_error_codes="True"
